[7.x] Split testing component so we can reused it on testing-browser-kit, lumen-framework etc.
