+++ src/Illuminate/Testing/Constraints/HasInDatabase.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Illuminate\Foundation\Testing\Constraints;
+namespace Illuminate\Testing\Constraints;
 
 use Illuminate\Database\Connection;
 use PHPUnit\Framework\Constraint\Constraint;
 
-class SoftDeletedInDatabase extends Constraint
+class HasInDatabase extends Constraint
 {
     /**
      * Number of records that will be shown in the console in case of failure.
@@ -29,27 +29,17 @@ class SoftDeletedInDatabase extends Constraint
     protected $data;
 
     /**
-     * The name of the column that indicates soft deletion has occurred.
-     *
-     * @var string
-     */
-    protected $deletedAtColumn;
-
-    /**
      * Create a new constraint instance.
      *
      * @param  \Illuminate\Database\Connection  $database
      * @param  array  $data
-     * @param  string  $deletedAtColumn
      * @return void
      */
-    public function __construct(Connection $database, array $data, string $deletedAtColumn)
+    public function __construct(Connection $database, array $data)
     {
         $this->data = $data;
 
         $this->database = $database;
-
-        $this->deletedAtColumn = $deletedAtColumn;
     }
 
     /**
@@ -60,10 +50,7 @@ class SoftDeletedInDatabase extends Constraint
      */
     public function matches($table): bool
     {
-        return $this->database->table($table)
-                ->where($this->data)
-                ->whereNotNull($this->deletedAtColumn)
-                ->count() > 0;
+        return $this->database->table($table)->where($this->data)->count() > 0;
     }
 
     /**
@@ -75,8 +62,8 @@ class SoftDeletedInDatabase extends Constraint
     public function failureDescription($table): string
     {
         return sprintf(
-            "any soft deleted row in the table [%s] matches the attributes %s.\n\n%s",
-            $table, $this->toString(), $this->getAdditionalInfo($table)
+            "a row in the table [%s] matches the attributes %s.\n\n%s",
+            $table, $this->toString(JSON_PRETTY_PRINT), $this->getAdditionalInfo($table)
         );
     }
 
@@ -108,10 +95,11 @@ class SoftDeletedInDatabase extends Constraint
     /**
      * Get a string representation of the object.
      *
+     * @param  int  $options
      * @return string
      */
-    public function toString(): string
+    public function toString($options = 0): string
     {
-        return json_encode($this->data);
+        return json_encode($this->data, $options);
     }
 }
