+++ src/Illuminate/Testing/Constraints/SoftDeletedInDatabase.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Illuminate\Foundation\Testing\Constraints;
+namespace Illuminate\Testing\Constraints;
 
 use Illuminate\Database\Connection;
 use PHPUnit\Framework\Constraint\Constraint;
 
-class HasInDatabase extends Constraint
+class SoftDeletedInDatabase extends Constraint
 {
     /**
      * Number of records that will be shown in the console in case of failure.
@@ -29,17 +29,27 @@ class HasInDatabase extends Constraint
     protected $data;
 
     /**
+     * The name of the column that indicates soft deletion has occurred.
+     *
+     * @var string
+     */
+    protected $deletedAtColumn;
+
+    /**
      * Create a new constraint instance.
      *
      * @param  \Illuminate\Database\Connection  $database
      * @param  array  $data
+     * @param  string  $deletedAtColumn
      * @return void
      */
-    public function __construct(Connection $database, array $data)
+    public function __construct(Connection $database, array $data, string $deletedAtColumn)
     {
         $this->data = $data;
 
         $this->database = $database;
+
+        $this->deletedAtColumn = $deletedAtColumn;
     }
 
     /**
@@ -50,7 +60,10 @@ class HasInDatabase extends Constraint
      */
     public function matches($table): bool
     {
-        return $this->database->table($table)->where($this->data)->count() > 0;
+        return $this->database->table($table)
+                ->where($this->data)
+                ->whereNotNull($this->deletedAtColumn)
+                ->count() > 0;
     }
 
     /**
@@ -62,8 +75,8 @@ class HasInDatabase extends Constraint
     public function failureDescription($table): string
     {
         return sprintf(
-            "a row in the table [%s] matches the attributes %s.\n\n%s",
-            $table, $this->toString(JSON_PRETTY_PRINT), $this->getAdditionalInfo($table)
+            "any soft deleted row in the table [%s] matches the attributes %s.\n\n%s",
+            $table, $this->toString(), $this->getAdditionalInfo($table)
         );
     }
 
@@ -95,11 +108,10 @@ class HasInDatabase extends Constraint
     /**
      * Get a string representation of the object.
      *
-     * @param  int  $options
      * @return string
      */
-    public function toString($options = 0): string
+    public function toString(): string
     {
-        return json_encode($this->data, $options);
+        return json_encode($this->data);
     }
 }
